improved documentation
renamed ActivityType to BehaviorMode
renamed ActivitySystem to BehaviorSystem
added MapToWorldConverter.java
    to make intent more clear in respect to WorldToMapConverter
Habitat.java
    food min / max constants are now all double values
EnvironmentDefinition.java
    deleted parameter simTime (obsolete with launcher)

+++ src/main/java/de/zmt/sim/Habitat.java
@@ -9,9 +9,18 @@ import org.jscience.physics.amount.Amount;
 import de.zmt.util.UnitConstants;
 import de.zmt.util.quantity.AreaDensity;
 
+/**
+ * Habitats available in the environment.
+ * <p>
+ * Minimum / maximum constants are estimated values based on information of
+ * algal standing crop. 5-14 g/m2 for coral reef as guideline.
+ * 
+ * @see "Cliffton 1995"
+ * 
+ * @author cmeyer
+ *
+ */
 public enum Habitat {
-    // algal turf standing crop in g/m2 as guideline:
-    // after Cliffton 1995 5-14 g/m2 in reef
     /** Coral reef: cyan color */
     CORALREEF {
 	@Override
@@ -207,109 +216,96 @@ public enum Habitat {
 
     private static final String CORALREEF_NAME = "coral reef";
     private static final Color CORALREEF_COLOR = Color.CYAN;
-    private static final int CORALREEF_FOOD_MIN_VALUE = 5;
-    private static final int CORALREEF_FOOD_MAX_VALUE = 14;
+    private static final double CORALREEF_FOOD_MIN_VALUE = 5;
+    private static final double CORALREEF_FOOD_MAX_VALUE = 14;
     private static final double CORALREEF_MORTALITY_RISK_PER_DAY_VALUE = 0.002;
 
     private static final String SEAGRASS_NAME = "seagrass bed";
     private static final Color SEAGRASS_COLOR = Color.GREEN;
-    private static final int SEAGRASS_FOOD_MIN_VALUE = 5;
-    private static final int SEAGRASS_FOOD_MAX_VALUE = 10;
+    private static final double SEAGRASS_FOOD_MIN_VALUE = 5;
+    private static final double SEAGRASS_FOOD_MAX_VALUE = 10;
     private static final double SEAGRASS_MORTALITY_RISK_PER_DAY_VALUE = 0.001;
 
     private static final String MANGROVE_NAME = "mangrove";
     private static final Color MANGROVE_COLOR = new Color(0, 178, 0);
-    private static final int MANGROVE_FOOD_MIN_VALUE = 3;
-    private static final int MANGROVE_FOOD_MAX_VALUE = 5;
+    private static final double MANGROVE_FOOD_MIN_VALUE = 3;
+    private static final double MANGROVE_FOOD_MAX_VALUE = 5;
     private static final double MANGROVE_MORTALITY_RISK_PER_DAY_VALUE = 0.002;
 
     private static final String ROCK_NAME = "rock";
     private static final Color ROCK_COLOR = Color.LIGHT_GRAY;
-    private static final int ROCK_FOOD_MIN_VALUE = 2;
-    private static final int ROCK_FOOD_MAX_VALUE = 5;
+    private static final double ROCK_FOOD_MIN_VALUE = 2;
+    private static final double ROCK_FOOD_MAX_VALUE = 5;
     private static final double ROCK_MORTALITY_RISK_PER_DAY_VALUE = 0.004;
 
     private static final String SANDYBOTTOM_NAME = "sandy bottom";
     private static final Color SANDYBOTTOM_COLOR = Color.YELLOW;
     private static final double SANDYBOTTOM_FOOD_MIN_VALUE = 0.1;
-    private static final int SANDYBOTTOM_FOOD_MAX_VALUE = 3;
+    private static final double SANDYBOTTOM_FOOD_MAX_VALUE = 3;
     private static final double SANDYBOTTOM_MORTALITY_RISK_PER_DAY_VALUE = 0.008;
 
     private static final String MAINLAND_NAME = "mainland";
     private static final Color MAINLAND_COLOR = Color.WHITE;
-    private static final int MAINLAND_FOOD_MIN_VALUE = 0;
-    private static final int MAINLAND_FOOD_MAX_VALUE = 0;
+    private static final double MAINLAND_FOOD_MIN_VALUE = 0;
+    private static final double MAINLAND_FOOD_MAX_VALUE = 0;
     private static final int MAINLAND_MORTALITY_RISK_PER_DAY_VALUE = 0;
 
-    private static final Amount<AreaDensity> CORALREEF_FOOD_MIN = Amount
-	    .valueOf(CORALREEF_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
-    private static final Amount<AreaDensity> CORALREEF_FOOD_MAX = Amount
-	    .valueOf(CORALREEF_FOOD_MAX_VALUE, UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> CORALREEF_FOOD_MIN = Amount.valueOf(CORALREEF_FOOD_MIN_VALUE,
+	    UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> CORALREEF_FOOD_MAX = Amount.valueOf(CORALREEF_FOOD_MAX_VALUE,
+	    UnitConstants.FOOD_DENSITY);
     private static final Amount<AreaDensity> CORALREEF_FOOD_RANGE = Amount
-	    .valueOf(CORALREEF_FOOD_MAX_VALUE - CORALREEF_FOOD_MIN_VALUE,
-		    UnitConstants.FOOD_DENSITY);
+	    .valueOf(CORALREEF_FOOD_MAX_VALUE - CORALREEF_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
     private static final Amount<Frequency> CORALREEF_MORTALITY_RISK = Amount
-	    .valueOf(CORALREEF_MORTALITY_RISK_PER_DAY_VALUE,
-		    UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
+	    .valueOf(CORALREEF_MORTALITY_RISK_PER_DAY_VALUE, UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
 
-    private static final Amount<AreaDensity> SEAGRASS_FOOD_MIN = Amount
-	    .valueOf(SEAGRASS_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
-    private static final Amount<AreaDensity> SEAGRASS_FOOD_MAX = Amount
-	    .valueOf(SEAGRASS_FOOD_MAX_VALUE, UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> SEAGRASS_FOOD_MIN = Amount.valueOf(SEAGRASS_FOOD_MIN_VALUE,
+	    UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> SEAGRASS_FOOD_MAX = Amount.valueOf(SEAGRASS_FOOD_MAX_VALUE,
+	    UnitConstants.FOOD_DENSITY);
     private static final Amount<AreaDensity> SEAGRASS_FOOD_RANGE = Amount
-	    .valueOf(SEAGRASS_FOOD_MAX_VALUE - SEAGRASS_FOOD_MIN_VALUE,
-		    UnitConstants.FOOD_DENSITY);
+	    .valueOf(SEAGRASS_FOOD_MAX_VALUE - SEAGRASS_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
     private static final Amount<Frequency> SEAGRASS_MORTALITY_RISK = Amount
-	    .valueOf(SEAGRASS_MORTALITY_RISK_PER_DAY_VALUE,
-		    UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
+	    .valueOf(SEAGRASS_MORTALITY_RISK_PER_DAY_VALUE, UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
 
-    private static final Amount<AreaDensity> MANGROVE_FOOD_MIN = Amount
-	    .valueOf(MANGROVE_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
-    private static final Amount<AreaDensity> MANGROVE_FOOD_MAX = Amount
-	    .valueOf(MANGROVE_FOOD_MAX_VALUE, UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> MANGROVE_FOOD_MIN = Amount.valueOf(MANGROVE_FOOD_MIN_VALUE,
+	    UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> MANGROVE_FOOD_MAX = Amount.valueOf(MANGROVE_FOOD_MAX_VALUE,
+	    UnitConstants.FOOD_DENSITY);
     private static final Amount<AreaDensity> MANGROVE_FOOD_RANGE = Amount
-	    .valueOf(MANGROVE_FOOD_MAX_VALUE - MANGROVE_FOOD_MIN_VALUE,
-		    UnitConstants.FOOD_DENSITY);
+	    .valueOf(MANGROVE_FOOD_MAX_VALUE - MANGROVE_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
     private static final Amount<Frequency> MANGROVE_MORTALITY_RISK = Amount
-	    .valueOf(MANGROVE_MORTALITY_RISK_PER_DAY_VALUE,
-		    UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
+	    .valueOf(MANGROVE_MORTALITY_RISK_PER_DAY_VALUE, UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
 
-    private static final Amount<AreaDensity> ROCK_FOOD_MIN = Amount.valueOf(
-	    ROCK_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
-    private static final Amount<AreaDensity> ROCK_FOOD_MAX = Amount.valueOf(
-	    ROCK_FOOD_MAX_VALUE, UnitConstants.FOOD_DENSITY);
-    private static final Amount<AreaDensity> ROCK_FOOD_RANGE = Amount.valueOf(
-	    ROCK_FOOD_MAX_VALUE - ROCK_FOOD_MIN_VALUE,
+    private static final Amount<AreaDensity> ROCK_FOOD_MIN = Amount.valueOf(ROCK_FOOD_MIN_VALUE,
+	    UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> ROCK_FOOD_MAX = Amount.valueOf(ROCK_FOOD_MAX_VALUE,
+	    UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> ROCK_FOOD_RANGE = Amount.valueOf(ROCK_FOOD_MAX_VALUE - ROCK_FOOD_MIN_VALUE,
 	    UnitConstants.FOOD_DENSITY);
     private static final Amount<Frequency> ROCK_MORTALITY_RISK = Amount
-	    .valueOf(ROCK_MORTALITY_RISK_PER_DAY_VALUE, UnitConstants.PER_DAY)
-	    .to(UnitConstants.PER_STEP);
+	    .valueOf(ROCK_MORTALITY_RISK_PER_DAY_VALUE, UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
 
-    private static final Amount<AreaDensity> SANDYBOTTOM_FOOD_MIN = Amount
-	    .valueOf(SANDYBOTTOM_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
-    private static final Amount<AreaDensity> SANDYBOTTOM_FOOD_MAX = Amount
-	    .valueOf(SANDYBOTTOM_FOOD_MAX_VALUE, UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> SANDYBOTTOM_FOOD_MIN = Amount.valueOf(SANDYBOTTOM_FOOD_MIN_VALUE,
+	    UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> SANDYBOTTOM_FOOD_MAX = Amount.valueOf(SANDYBOTTOM_FOOD_MAX_VALUE,
+	    UnitConstants.FOOD_DENSITY);
     private static final Amount<AreaDensity> SANDYBOTTOM_FOOD_RANGE = Amount
-	    .valueOf(SANDYBOTTOM_FOOD_MAX_VALUE - SANDYBOTTOM_FOOD_MIN_VALUE,
-		    UnitConstants.FOOD_DENSITY);
+	    .valueOf(SANDYBOTTOM_FOOD_MAX_VALUE - SANDYBOTTOM_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
     private static final Amount<Frequency> SANDYBOTTOM_MORTALITY_RISK = Amount
-	    .valueOf(SANDYBOTTOM_MORTALITY_RISK_PER_DAY_VALUE,
-		    UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
+	    .valueOf(SANDYBOTTOM_MORTALITY_RISK_PER_DAY_VALUE, UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
 
-    private static final Amount<AreaDensity> MAINLAND_FOOD_MIN = Amount
-	    .valueOf(MAINLAND_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
-    private static final Amount<AreaDensity> MAINLAND_FOOD_MAX = Amount
-	    .valueOf(MAINLAND_FOOD_MAX_VALUE, UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> MAINLAND_FOOD_MIN = Amount.valueOf(MAINLAND_FOOD_MIN_VALUE,
+	    UnitConstants.FOOD_DENSITY);
+    private static final Amount<AreaDensity> MAINLAND_FOOD_MAX = Amount.valueOf(MAINLAND_FOOD_MAX_VALUE,
+	    UnitConstants.FOOD_DENSITY);
     private static final Amount<AreaDensity> MAINLAND_FOOD_RANGE = Amount
-	    .valueOf(MAINLAND_FOOD_MAX_VALUE - MAINLAND_FOOD_MIN_VALUE,
-		    UnitConstants.FOOD_DENSITY);
+	    .valueOf(MAINLAND_FOOD_MAX_VALUE - MAINLAND_FOOD_MIN_VALUE, UnitConstants.FOOD_DENSITY);
     private static final Amount<Frequency> MAINLAND_MORTALITY_RISK = Amount
-	    .valueOf(MAINLAND_MORTALITY_RISK_PER_DAY_VALUE,
-		    UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
+	    .valueOf(MAINLAND_MORTALITY_RISK_PER_DAY_VALUE, UnitConstants.PER_DAY).to(UnitConstants.PER_STEP);
 
     public static Habitat DEFAULT = SANDYBOTTOM;
-    public static final double FOOD_RANGE_MAX = SEAGRASS_FOOD_MAX_VALUE
-	    - SEAGRASS_FOOD_MIN_VALUE;
+    public static final double FOOD_RANGE_MAX = SEAGRASS_FOOD_MAX_VALUE - SEAGRASS_FOOD_MIN_VALUE;
 
     /**
      * 
@@ -322,8 +318,7 @@ public enum Habitat {
 		return habitat;
 	    }
 	}
-	throw new IllegalArgumentException(color + " is not associated with a "
-		+ Habitat.class.getSimpleName());
+	throw new IllegalArgumentException(color + " is not associated with a " + Habitat.class.getSimpleName());
     }
 
     public abstract String getName();
@@ -335,10 +330,14 @@ public enum Habitat {
      */
     public abstract Color getColor();
 
-    /** @return Minimum food density in gram dry mass per square meter. */
+    /**
+     * @return Minimum food density in gram dry mass per square meter.
+     */
     public abstract Amount<AreaDensity> getFoodDensityMin();
 
-    /** @return Maximum food density in gram dry mass per square meter. */
+    /**
+     * @return Maximum food density in gram dry mass per square meter.
+     */
     public abstract Amount<AreaDensity> getFoodDensityMax();
 
     /**
@@ -347,6 +346,11 @@ public enum Habitat {
      */
     public abstract Amount<AreaDensity> getFoodDensityRange();
 
-    /** @return mortality risk for this habitat */
+    /**
+     * Estimated mortality risk as a summarizing factor of habitat complexity,
+     * available refuge and predator abundances.
+     * 
+     * @return mortality risk for this habitat
+     */
     public abstract Amount<Frequency> getMortalityRisk();
 }
